feat: moved exam generation functions to services where they are used
